Promote dual-stack listen into QUIC endpoint API
